+++ src/main/java/de/codescape/jira/plugins/multiplesubtasks/service/SubtaskTemplateService.java
@@ -2,8 +2,11 @@ package de.codescape.jira.plugins.multiplesubtasks.service;
 
 import com.atlassian.activeobjects.external.ActiveObjects;
 import com.atlassian.activeobjects.tx.Transactional;
+import com.atlassian.jira.project.Project;
+import com.atlassian.jira.user.ApplicationUser;
 import com.atlassian.plugin.spring.scanner.annotation.imports.ComponentImport;
 import de.codescape.jira.plugins.multiplesubtasks.ao.SubtaskConfig;
+import de.codescape.jira.plugins.multiplesubtasks.ao.SubtaskShare;
 import de.codescape.jira.plugins.multiplesubtasks.ao.SubtaskTemplate;
 import de.codescape.jira.plugins.multiplesubtasks.model.SubtaskTemplateType;
 import de.codescape.jira.plugins.multiplesubtasks.model.TemplateSortOrder;
@@ -13,7 +16,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 /**
  * Subtask template service to persist and load templates from and to the database.
@@ -35,9 +42,10 @@ public class SubtaskTemplateService {
     /**
      * Save a user specific subtask template.
      */
-    public void saveUserTemplate(Long userId, Long templateId, String templateName, String templateText) {
+    public SubtaskTemplate saveUserTemplate(Long userId, Long templateId, String templateName, String templateText) {
+        SubtaskTemplate subtaskTemplate;
         if (templateId == null) {
-            SubtaskTemplate subtaskTemplate = activeObjects.create(SubtaskTemplate.class,
+            subtaskTemplate = activeObjects.create(SubtaskTemplate.class,
                 new DBParam("TEMPLATE_TYPE", SubtaskTemplateType.USER),
                 new DBParam("USER_ID", userId),
                 new DBParam("NAME", templateName),
@@ -45,24 +53,34 @@ public class SubtaskTemplateService {
             );
             subtaskTemplate.save();
         } else {
-            SubtaskTemplate subtaskTemplate = findUserTemplate(userId, templateId);
+            subtaskTemplate = findUserTemplate(userId, templateId);
             if (subtaskTemplate != null) {
                 subtaskTemplate.setName(templateName);
                 subtaskTemplate.setTemplate(templateText);
                 subtaskTemplate.save();
             }
         }
+        return subtaskTemplate;
     }
 
     /**
      * Return a list of all subtask templates for a user.
      */
-    public List<SubtaskTemplate> getUserTemplates(Long userId) {
-        return Arrays.asList(activeObjects.find(SubtaskTemplate.class,
+    public List<SubtaskTemplate> getUserTemplates(Long userId, boolean includeSharedTemplates) {
+        List<SubtaskTemplate> ownTemplates = Arrays.asList(activeObjects.find(SubtaskTemplate.class,
             Query.select()
                 .where("USER_ID = ? and TEMPLATE_TYPE = ?", userId, SubtaskTemplateType.USER)
-                .order(getTemplateSortOrder()))
-        );
+                .order(getTemplateSortOrder())));
+        List<SubtaskTemplate> sharedTemplates = Collections.emptyList();
+        if (includeSharedTemplates) {
+            sharedTemplates = Arrays.asList(activeObjects.find(SubtaskTemplate.class,
+                Query.select()
+                    .alias(SubtaskTemplate.class, "ST")
+                    .alias(SubtaskShare.class, "SS")
+                    .join(SubtaskShare.class, "ST.ID = SS.TEMPLATE_ID")
+                    .where("SS.USER_ID = ? and ST.TEMPLATE_TYPE = ?", userId, SubtaskTemplateType.USER)));
+        }
+        return Stream.concat(ownTemplates.stream(), sharedTemplates.stream()).collect(Collectors.toList());
     }
 
     /**
@@ -71,6 +89,7 @@ public class SubtaskTemplateService {
     public void deleteUserTemplate(Long userId, Long templateId) {
         SubtaskTemplate subtaskTemplate = findUserTemplate(userId, templateId);
         if (subtaskTemplate != null) {
+            Arrays.stream(subtaskTemplate.getShares()).forEach(activeObjects::delete);
             activeObjects.delete(subtaskTemplate);
         }
     }
@@ -88,9 +107,10 @@ public class SubtaskTemplateService {
     /**
      * Save a project specific subtask template.
      */
-    public void saveProjectTemplate(Long projectId, Long userId, Long templateId, String templateName, String templateText) {
+    public SubtaskTemplate saveProjectTemplate(Long projectId, Long userId, Long templateId, String templateName, String templateText) {
+        SubtaskTemplate subtaskTemplate;
         if (templateId == null) {
-            SubtaskTemplate subtaskTemplate = activeObjects.create(SubtaskTemplate.class,
+            subtaskTemplate = activeObjects.create(SubtaskTemplate.class,
                 new DBParam("TEMPLATE_TYPE", SubtaskTemplateType.PROJECT),
                 new DBParam("PROJECT_ID", projectId),
                 new DBParam("USER_ID", userId),
@@ -99,7 +119,7 @@ public class SubtaskTemplateService {
             );
             subtaskTemplate.save();
         } else {
-            SubtaskTemplate subtaskTemplate = findProjectTemplate(projectId, templateId);
+            subtaskTemplate = findProjectTemplate(projectId, templateId);
             if (subtaskTemplate != null) {
                 subtaskTemplate.setUserId(userId);
                 subtaskTemplate.setName(templateName);
@@ -107,17 +127,27 @@ public class SubtaskTemplateService {
                 subtaskTemplate.save();
             }
         }
+        return subtaskTemplate;
     }
 
     /**
      * Return a list of all subtask templates for a project.
      */
-    public List<SubtaskTemplate> getProjectTemplates(Long projectId) {
-        return Arrays.asList(activeObjects.find(SubtaskTemplate.class,
+    public List<SubtaskTemplate> getProjectTemplates(Long projectId, boolean includeSharedTemplates) {
+        List<SubtaskTemplate> ownTemplates = Arrays.asList(activeObjects.find(SubtaskTemplate.class,
             Query.select()
                 .where("PROJECT_ID = ? and TEMPLATE_TYPE = ?", projectId, SubtaskTemplateType.PROJECT)
-                .order(getTemplateSortOrder()))
-        );
+                .order(getTemplateSortOrder())));
+        List<SubtaskTemplate> sharedTemplates = Collections.emptyList();
+        if (includeSharedTemplates) {
+            sharedTemplates = Arrays.asList(activeObjects.find(SubtaskTemplate.class,
+                Query.select()
+                    .alias(SubtaskTemplate.class, "ST")
+                    .alias(SubtaskShare.class, "SS")
+                    .join(SubtaskShare.class, "ST.ID = SS.TEMPLATE_ID")
+                    .where("SS.PROJECT_ID = ? and ST.TEMPLATE_TYPE = ?", projectId, SubtaskTemplateType.PROJECT)));
+        }
+        return Stream.concat(ownTemplates.stream(), sharedTemplates.stream()).collect(Collectors.toList());
     }
 
     /**
@@ -126,6 +156,7 @@ public class SubtaskTemplateService {
     public void deleteProjectTemplate(Long projectId, Long templateId) {
         SubtaskTemplate subtaskTemplate = findProjectTemplate(projectId, templateId);
         if (subtaskTemplate != null) {
+            Arrays.stream(subtaskTemplate.getShares()).forEach(activeObjects::delete);
             activeObjects.delete(subtaskTemplate);
         }
     }
@@ -152,9 +183,10 @@ public class SubtaskTemplateService {
     /**
      * Save a global subtask template.
      */
-    public void saveGlobalTemplate(Long userId, Long templateId, String templateName, String templateText) {
+    public SubtaskTemplate saveGlobalTemplate(Long userId, Long templateId, String templateName, String templateText) {
+        SubtaskTemplate subtaskTemplate;
         if (templateId == null) {
-            SubtaskTemplate subtaskTemplate = activeObjects.create(SubtaskTemplate.class,
+            subtaskTemplate = activeObjects.create(SubtaskTemplate.class,
                 new DBParam("TEMPLATE_TYPE", SubtaskTemplateType.GLOBAL),
                 new DBParam("USER_ID", userId),
                 new DBParam("NAME", templateName),
@@ -162,7 +194,7 @@ public class SubtaskTemplateService {
             );
             subtaskTemplate.save();
         } else {
-            SubtaskTemplate subtaskTemplate = findGlobalTemplate(templateId);
+            subtaskTemplate = findGlobalTemplate(templateId);
             if (subtaskTemplate != null) {
                 subtaskTemplate.setUserId(userId);
                 subtaskTemplate.setName(templateName);
@@ -170,6 +202,7 @@ public class SubtaskTemplateService {
                 subtaskTemplate.save();
             }
         }
+        return subtaskTemplate;
     }
 
     /**
@@ -188,6 +221,7 @@ public class SubtaskTemplateService {
     public void deleteGlobalTemplate(Long templateId) {
         SubtaskTemplate subtaskTemplate = findGlobalTemplate(templateId);
         if (subtaskTemplate != null) {
+            Arrays.stream(subtaskTemplate.getShares()).forEach(activeObjects::delete);
             activeObjects.delete(subtaskTemplate);
         }
     }
@@ -201,6 +235,40 @@ public class SubtaskTemplateService {
                 .where("TEMPLATE_TYPE = ?", SubtaskTemplateType.GLOBAL)
                 .order(getTemplateSortOrder()))
         );
+    }
+
+    /**
+     * Create a share for a given user template with a given user.
+     * <p>
+     * A template can never be shared with the template owner and not with the same user more than one time.
+     */
+    public void shareUserTemplate(SubtaskTemplate template, ApplicationUser user) {
+        SubtaskShare[] subtaskShares = activeObjects.find(SubtaskShare.class,
+            Query.select()
+                .where("TEMPLATE_ID = ? and USER_ID = ?", template.getID(), user.getId()));
+        if (!Objects.equals(template.getUserId(), user.getId()) && subtaskShares.length == 0) {
+            SubtaskShare subtaskShare = activeObjects.create(SubtaskShare.class,
+                new DBParam("TEMPLATE_ID", template.getID()),
+                new DBParam("USER_ID", user.getId()));
+            subtaskShare.save();
+        }
+    }
+
+    /**
+     * Create a share for a given project template with a given project.
+     * <p>
+     * A template can never be shared with the owning project and not with the same project more than one time.
+     */
+    public void shareProjectTemplate(SubtaskTemplate template, Project project) {
+        SubtaskShare[] subtaskShares = activeObjects.find(SubtaskShare.class,
+            Query.select()
+                .where("TEMPLATE_ID = ? and PROJECT_ID = ?", template.getID(), project.getId()));
+        if (!Objects.equals(template.getProjectId(), project.getId()) && subtaskShares.length == 0) {
+            SubtaskShare subtaskShare = activeObjects.create(SubtaskShare.class,
+                new DBParam("TEMPLATE_ID", template.getID()),
+                new DBParam("PROJECT_ID", project.getId()));
+            subtaskShare.save();
+        }
     }
 
 }