+++ src/test/java/de/codescape/jira/plugins/multiplesubtasks/MultipleSubtasksDatabaseUpdater.java
@@ -1,6 +1,7 @@
 package de.codescape.jira.plugins.multiplesubtasks;
 
 import de.codescape.jira.plugins.multiplesubtasks.ao.SubtaskConfig;
+import de.codescape.jira.plugins.multiplesubtasks.ao.SubtaskShare;
 import de.codescape.jira.plugins.multiplesubtasks.ao.SubtaskTemplate;
 import net.java.ao.EntityManager;
 import net.java.ao.test.jdbc.DatabaseUpdater;
@@ -13,7 +14,7 @@ public class MultipleSubtasksDatabaseUpdater implements DatabaseUpdater {
     @Override
     @SuppressWarnings("unchecked")
     public void update(EntityManager entityManager) throws Exception {
-        entityManager.migrate(SubtaskTemplate.class, SubtaskConfig.class);
+        entityManager.migrate(SubtaskTemplate.class, SubtaskConfig.class, SubtaskShare.class);
     }
 
 }